Extract anchored byte publisher
